I changed the input fields design because the previous design for some reason was having some inconsistency on different browsers
