+++ styles/globals.css
@@ -12,3 +12,16 @@
 .spinner-modal {
 	background-color: rgba(93, 103, 112, 0.774);
 }
+
+input[type="email"] {
+	box-shadow: 0px 0px 3px 1px black;
+	outline: none !important;
+}
+
+.password-host {
+	box-shadow: 0px 0px 3px 1px black;
+}
+
+.password-host input {
+	outline: none !important;
+}
